+++ src/WorkspaceDependencyViewExtension/WorkspaceDependencyViewExtension.cs
@@ -1,5 +1,7 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.Windows;
 using System.Windows.Controls;
 using Dynamo.Core;
 using Dynamo.Extensions;
@@ -8,6 +10,7 @@ using Dynamo.Logging;
 using Dynamo.PackageManager;
 using Dynamo.WorkspaceDependency.Properties;
 using Dynamo.Wpf.Extensions;
+using DynamoUtilities;
 
 namespace Dynamo.WorkspaceDependency
 {
@@ -20,12 +23,21 @@ namespace Dynamo.WorkspaceDependency
     {
         internal MenuItem workspaceReferencesMenuItem;
         private readonly String extensionName = Properties.Resources.ExtensionName;
+        private readonly string customNodeExtension = ".dyf";
 
         internal WorkspaceDependencyView DependencyView
         {
             get;
             set;
         }
+
+        /// <summary>
+        /// Internal cache of the data displayed in data grid, useful in unit testing.
+        /// You are not expected to modify this but rather inspection.
+        /// </summary>
+        internal IEnumerable<PackageDependencyRow> dataRows;
+        internal IEnumerable<DependencyRow> localDefinitionDataRows;
+        internal IEnumerable<DependencyRow> externalFilesDataRows;
 
         internal PackageManagerExtension pmExtension;
 
@@ -57,6 +69,9 @@ namespace Dynamo.WorkspaceDependency
         public override void Dispose()
         {
             DependencyView.Dispose();
+            dataRows = null;
+            localDefinitionDataRows = null;
+            externalFilesDataRows = null;
         }
 
 
@@ -77,9 +92,32 @@ namespace Dynamo.WorkspaceDependency
             this.MessageLogged?.Invoke(msg);
         }
 
+        private Boolean hasDependencyIssue = false;
+        /// <summary>
+        /// Property to raise to author's attention, if the Dynamo active workspace has any missing dependencies.
+        /// This determines if the package dep viewer will be injected into right panel.
+        /// </summary>
+        private Boolean HasDependencyIssue
+        {
+            get { return hasDependencyIssue; }
+            set
+            {
+                hasDependencyIssue = value;
+                if (hasDependencyIssue)
+                {
+                    DependencyView.loadedParams.AddToExtensionsSideBar(this, DependencyView);
+                    if (workspaceReferencesMenuItem != null && !workspaceReferencesMenuItem.IsChecked)
+                    {
+                        workspaceReferencesMenuItem.IsChecked = true;
+                    }
+                }
+            }
+        }
+
         public override void Loaded(ViewLoadedParams viewLoadedParams)
         {
             DependencyView = new WorkspaceDependencyView(this, viewLoadedParams);
+            DependencyRegen(DependencyView.currentWorkspace);
             // when a package is loaded update the DependencyView 
             // as we may have installed a missing package.
 
@@ -88,7 +126,7 @@ namespace Dynamo.WorkspaceDependency
             {
                 pmExtension.PackageLoader.PackgeLoaded += (package) =>
                 {
-                    DependencyView.DependencyRegen(viewLoadedParams.CurrentWorkspaceModel as WorkspaceModel, true);
+                    DependencyRegen(viewLoadedParams.CurrentWorkspaceModel as WorkspaceModel, true);
                 };
             }
 
@@ -99,7 +137,7 @@ namespace Dynamo.WorkspaceDependency
                 if (workspaceReferencesMenuItem.IsChecked)
                 {
                     // Refresh dependency data
-                    DependencyView.DependencyRegen(viewLoadedParams.CurrentWorkspaceModel as WorkspaceModel, true);
+                    DependencyRegen(viewLoadedParams.CurrentWorkspaceModel as WorkspaceModel, true);
                     viewLoadedParams.AddToExtensionsSideBar(this, DependencyView);
                     workspaceReferencesMenuItem.IsChecked = true;
                 }
@@ -110,6 +148,115 @@ namespace Dynamo.WorkspaceDependency
                 }
             };
             viewLoadedParams.AddExtensionMenuItem(workspaceReferencesMenuItem);
+        }
+
+                /// <summary>
+        /// Regenerate dependency table
+        /// </summary>
+        /// <param name="ws">workspace model</param>
+        /// <param name="forceCompute">flag indicating if the workspace references should be computed</param>
+        internal void DependencyRegen(WorkspaceModel ws, bool forceCompute = false)
+        {
+            DependencyView.RestartBanner.Visibility = Visibility.Hidden;
+            ws.ForceComputeWorkspaceReferences = forceCompute;
+
+            var packageDependencies = ws.NodeLibraryDependencies?.Where(d => d is PackageDependencyInfo).ToList();
+            var localDefinitions = ws.NodeLocalDefinitions?.Where(d => d is DependencyInfo).ToList();
+            var externalFiles = ws.ExternalFiles?.Where(d => d is DependencyInfo).ToList();
+
+            foreach (DependencyInfo info in localDefinitions)
+            {
+                try
+                {
+                    if (info.ReferenceType == ReferenceType.DYFFile)
+                    {
+                        // Try to get the Custom node information if possible.
+                        string customNodeName = info.Name.Replace(customNodeExtension, "");
+                        DependencyView.CustomNodeManager.TryGetNodeInfo(customNodeName, out CustomNodeInfo customNodeInfo);
+
+                        if (customNodeInfo != null)
+                        {
+                            info.Path = customNodeInfo.Path;
+                        }
+                    }
+                    info.Size = PathHelper.GetFileSize(info.Path);
+                }
+                catch (Exception ex)
+                {
+                    OnMessageLogged(LogMessage.Info(string.Format(Properties.Resources.DependencyViewExtensionErrorTemplate, ex.ToString())));
+                }
+
+                HasDependencyIssue = string.IsNullOrEmpty(info.Path);
+            }
+
+            foreach (DependencyInfo info in externalFiles)
+            {
+                HasDependencyIssue = string.IsNullOrEmpty(info.Path);
+            }
+
+            var pythonPackageDependencies = ws.OnRequestPackageDependencies();
+            if (pythonPackageDependencies != null)
+                packageDependencies.AddRange(pythonPackageDependencies);
+
+            if (packageDependencies.Any(d => d.State != PackageDependencyState.Loaded))
+            {
+                HasDependencyIssue = true;
+            }
+
+            if (packageDependencies.Any())
+            {
+                Boolean hasPackageMarkedForUninstall = false;
+                // If package is set to uninstall state, update the package info
+                foreach (var package in pmExtension.PackageLoader.LocalPackages.Where(x => 
+                x.LoadState.ScheduledState == PackageLoadState.ScheduledTypes.ScheduledForDeletion || x.LoadState.ScheduledState == PackageLoadState.ScheduledTypes.ScheduledForUnload))
+                {
+                    try
+                    {
+                        (packageDependencies.FirstOrDefault(x => x.Name == package.Name) as PackageDependencyInfo).State =
+                        PackageDependencyState.RequiresRestart;
+                        hasPackageMarkedForUninstall = true;
+                    }
+                    catch (Exception ex)
+                    {
+                        OnMessageLogged(LogMessage.Info(string.Format(Properties.Resources.DependencyViewExtensionErrorTemplate, $"failure to set package uninstall state |{ ex.ToString()}")));
+                    }
+                }
+
+                DependencyView.RestartBanner.Visibility = hasPackageMarkedForUninstall ? Visibility.Visible: Visibility.Hidden;
+            }
+
+            if (pmExtension != null)
+            {
+                foreach (PackageDependencyInfo packageDependencyInfo in packageDependencies)
+                {
+                    try
+                    {
+                        var targetInfo = pmExtension.PackageLoader.LocalPackages.Where(x => x.Name == packageDependencyInfo.Name).FirstOrDefault();
+                        if (targetInfo != null)
+                        {
+                            packageDependencyInfo.Path = targetInfo.RootDirectory;
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        OnMessageLogged(LogMessage.Info(string.Format(Properties.Resources.DependencyViewExtensionErrorTemplate, ex.ToString())));
+                    }
+                }
+            }
+
+            dataRows = packageDependencies.Select(d => new PackageDependencyRow(d as PackageDependencyInfo));
+            localDefinitionDataRows = localDefinitions.Select(d => new DependencyRow(d as DependencyInfo));
+            externalFilesDataRows = externalFiles.Select(d => new DependencyRow(d as DependencyInfo));
+
+            DependencyView.Packages.IsExpanded = dataRows.Count() > 0;
+            DependencyView.LocalDefinitions.IsExpanded = localDefinitionDataRows.Count() > 0;
+            DependencyView.ExternalFiles.IsExpanded = externalFilesDataRows.Count() > 0;
+
+            ws.ForceComputeWorkspaceReferences = false;
+
+            DependencyView.PackageDependencyTable.ItemsSource = dataRows;
+            DependencyView.LocalDefinitionsTable.ItemsSource = localDefinitionDataRows;
+            DependencyView.ExternalFilesTable.ItemsSource = externalFilesDataRows;
         }
 
         public override void Closed()
