+++ test/DynamoCoreWpfTests/ViewExtensions/WorkspaceDependencyViewExtensionTests.cs
@@ -66,7 +66,7 @@ namespace DynamoCoreWpfTests
             viewExtension.Loaded(loadedParams);
 
             var CurrentWorkspace = ViewModel.Model.CurrentWorkspace;
-            viewExtension.DependencyView.DependencyRegen(CurrentWorkspace);
+            viewExtension.DependencyRegen(CurrentWorkspace);
             // Restart banner should not display by default
             Assert.AreEqual(Visibility.Hidden, viewExtension.DependencyView.RestartBanner.Visibility);
         }
@@ -92,12 +92,12 @@ namespace DynamoCoreWpfTests
             package.LoadState.SetScheduledForDeletion();
 
             // Once choosing to install the specified version, info.State should reflect RequireRestart
-            viewExtension.DependencyView.DependencyRegen(CurrentWorkspace);
+            viewExtension.DependencyRegen(CurrentWorkspace);
 
             // Restart banner should display immediately
             Assert.AreEqual(Visibility.Visible, viewExtension.DependencyView.RestartBanner.Visibility);
             Assert.AreEqual(1, viewExtension.DependencyView.PackageDependencyTable.Items.Count);
-            var newInfo = viewExtension.DependencyView.dataRows.FirstOrDefault().DependencyInfo;
+            var newInfo = viewExtension.dataRows.FirstOrDefault().DependencyInfo;
 
             // Local loaded version was 2.0.0, but now will be update to date with dyn
             Assert.AreEqual("2.0.1", newInfo.Version.ToString());
@@ -214,7 +214,7 @@ namespace DynamoCoreWpfTests
             // Closing the dyf will trigger DependencyRegen of HomeWorkspaceModel.
             // The HomeWorkspaceModel does not contain any dependency info since it's empty
             // but DependencyRegen() call on it should not crash
-            Assert.DoesNotThrow(()=> viewExtension.DependencyView.DependencyRegen(homeWorkspaceModel));
+            Assert.DoesNotThrow(()=> viewExtension.DependencyRegen(homeWorkspaceModel));
         }
 
         [Test]
@@ -291,7 +291,7 @@ namespace DynamoCoreWpfTests
             Open(examplePath);
             Assert.AreEqual(1, ViewModel.SideBarTabItems.Count);
 
-            foreach (PackageDependencyRow packageDependencyRow in WorkspaceReferencesExtension.DependencyView.dataRows)
+            foreach (PackageDependencyRow packageDependencyRow in WorkspaceReferencesExtension.dataRows)
             {
                 var dependencyInfo = packageDependencyRow.DependencyInfo;
                 Assert.Contains(dependencyInfo.Name, dependenciesList);
@@ -314,8 +314,8 @@ namespace DynamoCoreWpfTests
             examplePath = Path.Combine(@"core\LocalDefinitionsTest.dyn");
             Open(examplePath);
            
-            Assert.AreEqual(1, WorkspaceReferencesExtension.DependencyView.localDefinitionDataRows.Count());
-            DependencyRow localDefinitionRow = WorkspaceReferencesExtension.DependencyView.localDefinitionDataRows.FirstOrDefault();
+            Assert.AreEqual(1, WorkspaceReferencesExtension.localDefinitionDataRows.Count());
+            DependencyRow localDefinitionRow = WorkspaceReferencesExtension.localDefinitionDataRows.FirstOrDefault();
             var dependencyInfo = localDefinitionRow.DependencyInfo;
             Assert.Contains(dependencyInfo.Name, dependenciesList);
         }
@@ -329,10 +329,10 @@ namespace DynamoCoreWpfTests
             var examplePath = Path.Combine(@"core\ExternalReferencesTest.dyn");
             Open(examplePath);
 
-            WorkspaceReferencesExtension.DependencyView.DependencyRegen(Model.CurrentWorkspace, true);
+            WorkspaceReferencesExtension.DependencyRegen(Model.CurrentWorkspace, true);
 
-            Assert.AreEqual(2, WorkspaceReferencesExtension.DependencyView.externalFilesDataRows.Count());
-            foreach (DependencyRow localDefinitionRow in WorkspaceReferencesExtension.DependencyView.externalFilesDataRows)
+            Assert.AreEqual(2, WorkspaceReferencesExtension.externalFilesDataRows.Count());
+            foreach (DependencyRow localDefinitionRow in WorkspaceReferencesExtension.externalFilesDataRows)
             {
                 var dependencyInfo = localDefinitionRow.DependencyInfo;
                 Assert.Contains(dependencyInfo.Name, dependenciesList);
@@ -346,11 +346,11 @@ namespace DynamoCoreWpfTests
             var examplePath = Path.Combine(@"core\ExternalReferencesTest.dyn");
             Open(examplePath);
             (Model.CurrentWorkspace as HomeWorkspaceModel).RunSettings.RunEnabled = false;
-            WorkspaceReferencesExtension.DependencyView.DependencyRegen(Model.CurrentWorkspace, true);
+            WorkspaceReferencesExtension.DependencyRegen(Model.CurrentWorkspace, true);
             var results = Model.CurrentWorkspace.ExternalFiles;
             Assert.AreEqual(0, results.Count());
             (Model.CurrentWorkspace as HomeWorkspaceModel).RunSettings.RunEnabled = true;
-            WorkspaceReferencesExtension.DependencyView.DependencyRegen(Model.CurrentWorkspace, true);
+            WorkspaceReferencesExtension.DependencyRegen(Model.CurrentWorkspace, true);
             results = Model.CurrentWorkspace.ExternalFiles;
             Assert.AreEqual(2, results.Count());
         }
